Fix PeriodicTimer to use nullptr for functions instead of the Null Pattern
This seems to avoid this weird behaviour where the 'this' of a lambda's
surrounding code changes if end() is called in the middle or at the top.
(end() replaces the current function with the nullptr/Null function.) It's
not clear why assignment would have this behaviour. Is it expected or is it
a compiler issue?